@@ -11,13 +11,13 @@
   "license": "MIT",
   "gypfile": true,
   "devDependencies": {
-    "electron": "1.8.4",
+    "electron": "9.0.3",
     "mocha": "^3.4.2",
     "shelljs": "^0.7.8",
     "should": "^11.2.1"
   },
   "dependencies": {
-    "nan": "^2.7.0"
+    "nan": "^2.14.1"
   },
   "engines": {
     "node": ">=8.2.1"
