@@ -33,8 +33,8 @@
     "commander": "~2.9.0",
     "cross-env": "^3.1.4",
     "del": "^4.1.1",
-    "electron": "1.8.4",
-    "electron-packager": "7.1.0",
+    "electron": "9.0.3",
+    "electron-packager": "^14.2.1",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.10.1",
     "eslint-plugin-prettier": "^3.1.2",
@@ -64,10 +64,10 @@
     "spectron": "^3.7.2"
   },
   "dependencies": {
-    "ali-oss": "~6.9.0",
+    "ali-oss": "^6.1.1",
     "aliyun-sdk": "^1.12.3",
     "clipboard": "^1.7.1",
-    "electron-log": "1.3.0",
+    "electron-log": "^4.2.1",
     "koa": "^2.3.0",
     "koa-convert": "^1.2.0",
     "koa-static-server": "^1.2.1",
