@@ -1,7 +1,10 @@
 angular.module("web").factory("Cipher", function () {
   var crypto = require("crypto");
-  var ALGORITHM = "aes192";
-  var KEY = "x82m#*lx8vv";
+  // 通过crypto.getCiphers()确认，aes192算法已不被支持
+  // todo: 历史数据的保存方案
+  var ALGORITHM = "aes-192-cbc";
+  var KEY = "x82m#*lx8vv0123456789abc";
+  var IV = "1234567890123456";
 
   return {
     cipher: cipher,
@@ -9,11 +12,11 @@ angular.module("web").factory("Cipher", function () {
   };
 
   function cipher(buf, key, algorithm) {
-    if (!buf instanceof Buffer) {
-      buf = new Buffer(buf);
+    if (!(buf instanceof Buffer)) {
+      buf = Buffer.from(buf);
     }
     var encrypted = "";
-    var cip = crypto.createCipher(algorithm || ALGORITHM, key || KEY);
+    var cip = crypto.createCipheriv(algorithm || ALGORITHM, key || KEY, IV);
     encrypted += cip.update(buf, "utf8", "hex");
     encrypted += cip.final("hex");
     return encrypted;
@@ -21,7 +24,11 @@ angular.module("web").factory("Cipher", function () {
 
   function decipher(encrypted, key, algorithm) {
     var decrypted = "";
-    var decipher = crypto.createDecipher(algorithm || ALGORITHM, key || KEY);
+    var decipher = crypto.createDecipheriv(
+      algorithm || ALGORITHM,
+      key || KEY,
+      IV
+    );
     decrypted += decipher.update(encrypted, "hex", "utf8");
     decrypted += decipher.final("utf8");
     return decrypted;
